+++ lib/testScenario/postmanCollectionRunnerClient.ts
@@ -1,4 +1,4 @@
-import path from "path";
+import path, { dirname } from "path";
 import newman from "newman";
 import {
   VariableScope,
@@ -19,6 +19,7 @@ import {
 import { inject, injectable } from "inversify";
 import { JsonLoader, JsonLoaderOption } from "../swagger/jsonLoader";
 import { setDefaultOpts } from "../swagger/loader";
+import { printWarning } from "../util/utils";
 import { ValidationLevel } from "./reportGenerator";
 import { SwaggerAnalyzer } from "./swaggerAnalyzer";
 import { DataMasker } from "./dataMasker";
@@ -45,6 +46,7 @@ import {
   defaultNewmanReport,
 } from "./defaultNaming";
 import { NewmanReport } from "./postmanReportParser";
+import { RuntimeEnvManager } from "./runtimeEnvManager";
 
 export interface PostmanCollectionRunnerClientOption extends BlobUploaderOption, JsonLoaderOption {
   testScenarioFileName: string;
@@ -60,6 +62,9 @@ export interface PostmanCollectionRunnerClientOption extends BlobUploaderOption,
   swaggerFilePaths?: string[];
   baseUrl: string;
   validationLevel?: ValidationLevel;
+  skipCleanUp?: boolean;
+  from?: string;
+  to?: string;
 }
 
 function makeid(length: number): string {
@@ -160,6 +165,9 @@ export class PostmanCollectionRunnerClient implements TestScenarioRunnerClient {
     subscriptionId: string,
     _resourceGroupName: string
   ): Promise<void> {
+    if (this.opts.from || this.opts.to) {
+      return;
+    }
     const item = new Item({
       name: "deleteResourceGroup",
       request: {
@@ -490,6 +498,26 @@ export class PostmanCollectionRunnerClient implements TestScenarioRunnerClient {
         this.opts.testScenarioName
       )}`
     );
+    const runtimeEnvManager = new RuntimeEnvManager(
+      path.join(dirname(reportExportPath), this.opts.testScenarioName),
+      this.opts,
+      this.collection
+    );
+
+    if (this.opts.from) {
+      const lastRnv = runtimeEnvManager.loadEnv(this.opts.from);
+      this.collectionEnv.syncVariablesFrom(lastRnv);
+      // use the variables value which exist in the env.json or process.env
+      for (const k of Object.keys(this.collectionEnv.syncVariablesTo())) {
+        const v = this.opts.env.get(k);
+        if (v) {
+          this.collectionEnv.set(k, v, typeof v);
+        }
+      }
+    }
+    if (this.opts.from || this.opts.to) {
+      runtimeEnvManager.repopulateCollectionItems(this.opts.from, this.opts.to);
+    }
     newman
       .run(
         {
@@ -508,6 +536,16 @@ export class PostmanCollectionRunnerClient implements TestScenarioRunnerClient {
           console.log("collection run complete!");
         }
       )
+      .on("beforeItem", async function (this: any, _err, _summary) {
+        if (!_err) {
+          runtimeEnvManager.save(_summary.item.name, this, "beforeStep");
+        }
+      })
+      .on("item", async function (this: any, _err, _summary) {
+        if (!_err) {
+          runtimeEnvManager.clean();
+        }
+      })
       .on("done", async (_err, _summary) => {
         const keys = await this.swaggerAnalyzer.getAllSecretKey();
         const values: string[] = [];
@@ -551,6 +589,13 @@ export class PostmanCollectionRunnerClient implements TestScenarioRunnerClient {
         };
         const reportAnalyzer = inversifyGetInstance(NewmanReportAnalyzer, opts);
         await reportAnalyzer.analyze();
+        if (this.opts.skipCleanUp || this.opts.to) {
+          printWarning(
+            `Notice:the resource group '${this.collectionEnv.get(
+              "resourceGroupName"
+            )}' was not cleaned up.`
+          );
+        }
       });
   }